JPA 전환 준비중 -> 기존 매퍼 제거, 도메인 일부 변경중

+++ src/main/java/com/vig/domain/ImageInfo.java
@@ -13,7 +13,7 @@ public class ImageInfo implements Serializable{
 	
 	private static final long serialVersionUID = 1236283424582915L;
 	
-	private int imageId;
+	private long imageId;
 	private String path;
 	
 }
